+++ x/tx/signing/textual/coin_test.go
@@ -52,18 +52,18 @@ func addMetadataToContext(ctx context.Context, metadata *bankv1beta1.Metadata) c
 
 func TestMetadataQuerier(t *testing.T) {
 	// Errors on nil metadata querier
-	txt := textual.NewSignModeHandler(nil)
-	vr, err := txt.GetFieldValueRenderer(fieldDescriptorFromName("COIN"))
-	require.NoError(t, err)
-	_, err = vr.Format(context.Background(), protoreflect.ValueOf((&basev1beta1.Coin{}).ProtoReflect()))
+	_, err := textual.NewSignModeHandler(textual.SignModeOptions{})
 	require.Error(t, err)
 
 	// Errors if metadata querier returns an error
 	expErr := fmt.Errorf("mock error")
-	txt = textual.NewSignModeHandler(func(_ context.Context, _ string) (*bankv1beta1.Metadata, error) {
-		return nil, expErr
+	txt, err := textual.NewSignModeHandler(textual.SignModeOptions{
+		CoinMetadataQuerier: func(_ context.Context, _ string) (*bankv1beta1.Metadata, error) {
+			return nil, expErr
+		},
 	})
-	vr, err = txt.GetFieldValueRenderer(fieldDescriptorFromName("COIN"))
+	require.NoError(t, err)
+	vr, err := txt.GetFieldValueRenderer(fieldDescriptorFromName("COIN"))
 	require.NoError(t, err)
 	_, err = vr.Format(context.Background(), protoreflect.ValueOf((&basev1beta1.Coin{}).ProtoReflect()))
 	require.ErrorIs(t, err, expErr)
@@ -78,7 +78,8 @@ func TestCoinJsonTestcases(t *testing.T) {
 	err = json.Unmarshal(raw, &testcases)
 	require.NoError(t, err)
 
-	textual := textual.NewSignModeHandler(mockCoinMetadataQuerier)
+	textual, err := textual.NewSignModeHandler(textual.SignModeOptions{CoinMetadataQuerier: mockCoinMetadataQuerier})
+	require.NoError(t, err)
 	vr, err := textual.GetFieldValueRenderer(fieldDescriptorFromName("COIN"))
 	require.NoError(t, err)