+++ x/tx/signing/textual/repeated_test.go
@@ -29,7 +29,7 @@ func TestRepeatedJsonTestcases(t *testing.T) {
 	err = json.Unmarshal(raw, &testcases)
 	require.NoError(t, err)
 
-	tr := textual.NewSignModeHandler(mockCoinMetadataQuerier)
+	tr, err := textual.NewSignModeHandler(textual.SignModeOptions{CoinMetadataQuerier: mockCoinMetadataQuerier})
 	for i, tc := range testcases {
 		t.Run(fmt.Sprintf("%d", i), func(t *testing.T) {
 			// Create a context.Context containing all coins metadata, to simulate